refactor: split cli.rs into per-subcommand modules
cli.rs held 1359 lines: the clap declarations for every subcommand, plus
logic that is not declarative at all, plus the tests. Each new subcommand
added its arguments inline to the Commands enum, so the file grew
monotonically.

Commands now holds an argument struct per variant, and each subcommand's
struct and action enum lives in its own module, so adding a subcommand
adds a file instead of extending a single 500-line enum. RArgsBuilder,
completion generation and the headless R-source precedence helper move out
too, since none of them are clap definitions.

Every match site that destructured the old struct-style variants was
updated for the tuple form. That work is mechanical, and deliberately kept
so: no handler's logic changed, only how it receives its arguments.

The generated completion and help snapshots are byte-identical, which is
what proves the CLI surface did not move. The test count is unchanged.
